Move search-results routes to components
They are no longer routes, so they shouldn't be with routes.

Since all of the directories inside of `src/routes` only had a single
file, I also moved those files out into the root of the routes directory.
