+++ src/components/title-search-results.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
 import Icon from '@folio/stripes-components/lib/Icon';
 
-import List from '../../components/list';
-import VendorListItem from '../../components/vendor-list-item';
+import List from './list';
+import TitleListItem from './title-list-item';
 
-function VendorSearchResultsRoute({
+export default function TitleSearchResults({
   location,
   query: { search },
   isPending,
@@ -18,30 +17,30 @@ function VendorSearchResultsRoute({
   return isPending ? (
     <Icon icon="spinner-ellipsis" />
   ) : isRejected ? (
-    <p data-test-vendor-search-error-message>
+    <p data-test-title-search-error-message>
       {error.length ? error[0].message : error.message}
     </p>
   ) : isResolved && !content.length ? (
-    <p data-test-vendor-search-no-results>
-      No vendors found for <strong>{`"${search}"`}</strong>.
+    <p data-test-title-search-no-results>
+      No titles found for <strong>{`"${search}"`}</strong>.
     </p>
   ) : (
-    <List data-test-vendor-search-results-list>
-      {content.map(vendor => (
-        <VendorListItem
-          key={vendor.vendorId}
-          item={vendor}
+    <List data-test-title-search-results-list>
+      {content.map(title => (
+        <TitleListItem
+          key={title.titleId}
           link={{
-            pathname: `/eholdings/vendors/${vendor.vendorId}`,
+            pathname: `/eholdings/titles/${title.titleId}`,
             search: location.search
           }}
+          item={title}
         />
       ))}
     </List>
   );
 }
 
-VendorSearchResultsRoute.propTypes = {
+TitleSearchResults.propTypes = {
   location: PropTypes.object.isRequired,
   query: PropTypes.shape({
     search: PropTypes.string
@@ -55,7 +54,3 @@ VendorSearchResultsRoute.propTypes = {
     PropTypes.array
   ])
 };
-
-export default connect(
-  ({ eholdings: { search } }) => ({ ...search.vendors })
-)(VendorSearchResultsRoute);
