+++ src/routes/title-show.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { getTitle } from '../../redux/title';
+import { getTitle } from '../redux/title';
 
-import View from '../../components/title-show';
+import View from '../components/title-show';
 
 class TitleShowRoute extends Component {
   static propTypes = {
